Simplify creation of routes
